requirements.txt
# Core dependencies
pipecat>=0.0.1
baml>=0.19.0,<1.0.0
openai>=1.0.0  # optional if using Gemini
google-generativeai>=0.7.2
python-dotenv>=1.0.0

# Audio processing (optional)
pyaudio>=0.2.11
sounddevice>=0.4.6

# Async support
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
